+++ kitty/cursor_trail.c
@@ -160,12 +160,9 @@ update_cursor_trail_needs_render(CursorTrail *ct, Window *w, ndc_coords g) {
 
 bool
 update_cursor_trail(CursorTrail *ct, Window *w, monotonic_t now, OSWindow *os_window) {
-    if (ct->has_previous_position) {
-        for (int i = 0; i < 4; i++) {
-            ct->previous_corner_x[i] = ct->corner_x[i];
-            ct->previous_corner_y[i] = ct->corner_y[i];
-        }
-    }
+    // remember where the trail was drawn in the previous frame, for shaders that do motion blur
+    memcpy(ct->previous_corner_x, ct->corner_x, sizeof(ct->previous_corner_x));
+    memcpy(ct->previous_corner_y, ct->corner_y, sizeof(ct->previous_corner_y));
     ct->target_updated = false;
     ndc_coords g = {
         .xstart = gl_pos_x(w->render_data.geometry.left, os_window->viewport_width),
@@ -184,14 +181,6 @@ update_cursor_trail(CursorTrail *ct, Window *w, monotonic_t now, OSWindow *os_wi
 
     bool needs_render_prev = ct->needs_render;
     update_cursor_trail_needs_render(ct, w, g);
-
-    if (!ct->has_previous_position) {
-        for (int i = 0; i < 4; i++) {
-            ct->previous_corner_x[i] = ct->corner_x[i];
-            ct->previous_corner_y[i] = ct->corner_y[i];
-        }
-        ct->has_previous_position = true;
-    }
 
     ct->updated_at = now;